Change EmojiIds type from byte[] to int[] in responses
Updated RegisterResponse.EmojiIds from byte[] to int[] for consistency. Adjusted Register class to cast emoji IDs to int when creating the response. This standardizes the representation of emoji IDs across the codebase.

+++ api/Models/RegisterRequest.cs
@@ -15,5 +15,5 @@ public sealed class RegisterRequest
 public sealed class RegisterResponse
 {
     public required Guid PairId { get; init;}
-    public required byte[] EmojiIds { get; init; } // length 4
+    public required int[] EmojiIds { get; init; } // length 4
 }
